Successful deploy of MiniMe token
Changed the source of the minime token. This source contains a
seperate file for the the MiniMe Factory. The reason for seperating
the solidity files is because the artifacts.require(<sol file>)
only imports a single contract. The old minime solidity file had
multiple contracts in it so it wasn't import the MiniMeFactory
Acontract. Also fixed AltheaDAO minime creatClone function parameters
to match the MimiMeToken function