Changed the data structure of users and bills
This was done because there wasn't a way for the Aragon app
to fetch the ipv6 address. The key change was changing the
variable type of the subnetSubcribers from address to bytes16.
This allows the aragon app to then query the users list (previously
the node list) and obtain the 2 following pieces of information,
an ethereum address and a nick name. The eth address is enough to
then find all the bill information it needs since that mapping
stays the same.

Another change is the combination of nodeList and nick name.
They where essentially always going to be the same person
so keeping them together made sense.